funční mini-client (work in progress)
