+++ src/cro/schedule/__init__.py
@@ -1,8 +1,7 @@
 # -*- coding: utf-8 -*-
 
-from cro.schedule._client import Client
 from cro.schedule._domain import DaySchedule, Show, Station
 
-__all__ = tuple(["Client", "Station", "Show", "DaySchedule"])
+__all__ = tuple(["Station", "Show", "DaySchedule", "Broadcast", "Person"])
 
-__version__ = "0.14.0" #????
+__version__ = "0.1.0" #????
